cleanup and move types to separate file
