+++ backend/protocol_rpc/message_handler/types.py
@@ -1,17 +1,33 @@
+from enum import Enum
 from dataclasses import dataclass
 
-from backend.protocol_rpc.types import EndpointResult
+
+class EventType(Enum):
+    DEBUG = "debug"
+    INFO = "info"
+    SUCCESS = "success"
+    ERROR = "error"
+
+
+class EventScope(Enum):
+    RPC = "RPC"
+    GENVM = "GenVM"
+    CONSENSUS = "Consensus"
 
 
 @dataclass
-class FormattedResponse:
-    function_name: str
-    trace_id: str
-    response: EndpointResult
+class LogEvent:
+    name: str
+    type: EventType
+    scope: EventScope
+    message: str
+    data: dict = None
 
-    def to_json(self):
+    def to_dict(self):
         return {
-            "function_name": self.function_name,
-            "trace_id": self.trace_id,
-            "response": self.response.to_json(),
+            "name": self.name,
+            "type": self.type.value,
+            "scope": self.scope.value,
+            "message": self.message,
+            "data": self.data,
         }
